refactor: extract duplicate buildFileContent logic to MetadataHelpers
- Extracted buildFileContent() from 3 creation services
- Added MetadataHelpers.buildFileContent() with optional bodyContent parameter
- Removed duplicate implementations (42 lines → 14 lines shared utility)
- Updated tests to use MetadataHelpers instead of private methods
- Impact: -28 lines duplication

Services affected:
- AreaCreationService: removed 13 lines
- ProjectCreationService: removed 13 lines
- TaskCreationService: removed 16 lines (includes JSDoc)
- MetadataHelpers: added 14 lines (shared utility)

Test updates:
- TaskCreationService.test.ts: Import MetadataHelpers, use static method
