+++ src/infrastructure/services/ProjectCreationService.ts
@@ -33,7 +33,7 @@ export class ProjectCreationService {
       label,
       uid,
     );
-    const fileContent = this.buildFileContent(frontmatter);
+    const fileContent = MetadataHelpers.buildFileContent(frontmatter);
 
     const folderPath = sourceFile.parent?.path || "";
     const filePath = folderPath ? `${folderPath}/${fileName}` : fileName;
@@ -78,19 +78,5 @@ export class ProjectCreationService {
     }
 
     return frontmatter;
-  }
-
-  private buildFileContent(frontmatter: Record<string, any>): string {
-    const frontmatterLines = Object.entries(frontmatter)
-      .map(([key, value]) => {
-        if (Array.isArray(value)) {
-          const arrayItems = value.map((item) => `  - ${item}`).join("\n");
-          return `${key}:\n${arrayItems}`;
-        }
-        return `${key}: ${value}`;
-      })
-      .join("\n");
-
-    return `---\n${frontmatterLines}\n---\n\n`;
   }
 }
